+++ database/migrations/2026_08_24_231624_create_detalle_pedidos_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('detalle_pedidos', function (Blueprint $table) {
-         $table->id();
-         $table->foreignId('pedido_id')->constrained('pedidos')->onDelete('cascade');
-         $table->foreignId('platillo_id')->constrained('platillos')->onDelete('cascade');
-         $table->integer('cantidad');
-         $table->decimal('precio_unitario', 10, 2);
-         $table->timestamps();
-        });
+        if (!Schema::hasTable('detalle_pedidos')) {
+            Schema::create('detalle_pedidos', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('pedido_id')->constrained('pedidos')->onDelete('cascade');
+                $table->foreignId('platillo_id')->constrained('platillos')->onDelete('cascade');
+                $table->integer('cantidad');
+                $table->decimal('precio_unitario', 10, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
